+++ src/index-shadow.ts
@@ -1,0 +1,6 @@
+import { polyfill } from './polyfill.js';
+import { installConstructedStylesheetPatches } from './shadow.js';
+
+installConstructedStylesheetPatches();
+
+export default polyfill;
