+++ src/shadow.ts
@@ -1,0 +1,88 @@
+import { polyfill } from './polyfill.js';
+import { captureAdoptedStylesheetText, originalReplaceSync } from './utils.js';
+
+interface CustomElementHost extends HTMLElement {
+  connectedCallback?: () => void;
+}
+
+// Marks host elements whose `connectedCallback` has already been wrapped, so we
+// don't wrap it more than once if multiple stylesheets are adopted.
+const patchedHosts = new WeakSet<HTMLElement>();
+
+/**
+ * Wraps the `connectedCallback` of a shadow root's host element so that, after
+ * the original callback runs (and the shadow DOM is populated), the polyfill is
+ * run for that shadow root to position its anchored elements.
+ */
+function patchHostConnectedCallback(shadowRoot: ShadowRoot) {
+  const host = shadowRoot.host as CustomElementHost;
+  if (patchedHosts.has(host)) {
+    return;
+  }
+  patchedHosts.add(host);
+
+  const originalConnectedCallback = host.connectedCallback;
+  host.connectedCallback = function (this: CustomElementHost) {
+    originalConnectedCallback?.call(this);
+    void polyfill({ roots: [shadowRoot] });
+  };
+
+  // If the host is already connected (e.g. `adoptedStyleSheets` was assigned
+  // from within the host's `connectedCallback`), the wrapper above won't run
+  // for the current connection, so run the polyfill once the current callback
+  // has finished and the shadow DOM has been populated.
+  if (host.isConnected) {
+    queueMicrotask(() => {
+      void polyfill({ roots: [shadowRoot] });
+    });
+  }
+}
+
+/**
+ * Installs patches on `CSSStyleSheet.prototype.replaceSync` and the
+ * `ShadowRoot.prototype.adoptedStyleSheets` setter to support CSS anchor
+ * positioning in constructed stylesheets adopted into shadow roots.
+ *
+ * Call this as early as possible — before any custom element's
+ * `connectedCallback` runs — so that constructed stylesheets are captured and
+ * their shadow roots are queued for positioning.
+ */
+export function installConstructedStylesheetPatches() {
+  // Patch `replaceSync` to capture the source text of constructed stylesheets
+  // so the polyfill can later re-parse it.
+  if (CSSStyleSheet.prototype.replaceSync === originalReplaceSync) {
+    CSSStyleSheet.prototype.replaceSync = function (text: string) {
+      captureAdoptedStylesheetText(this, text);
+      return originalReplaceSync.call(this, text);
+    };
+  }
+
+  const adoptedStyleSheetsDescriptor = Object.getOwnPropertyDescriptor(
+    ShadowRoot.prototype,
+    'adoptedStyleSheets',
+  );
+  const originalAdoptedStyleSheetsSet = adoptedStyleSheetsDescriptor?.set;
+
+  // Patch the `adoptedStyleSheets` setter so that anchors/targets styled by
+  // constructed stylesheets get positioned. We have access to the `ShadowRoot`
+  // here (`this`), but the shadow root's children may not exist yet (e.g. when
+  // `adoptedStyleSheets` is assigned before `innerHTML` in `connectedCallback`).
+  // To position only after the shadow DOM is populated, we wrap the host
+  // element's `connectedCallback` and run the polyfill for the shadow root once
+  // the (original) callback has finished.
+  if (
+    adoptedStyleSheetsDescriptor &&
+    originalAdoptedStyleSheetsSet &&
+    adoptedStyleSheetsDescriptor.set === originalAdoptedStyleSheetsSet
+  ) {
+    Object.defineProperty(ShadowRoot.prototype, 'adoptedStyleSheets', {
+      ...adoptedStyleSheetsDescriptor,
+      set(this: ShadowRoot, sheets: CSSStyleSheet[]) {
+        originalAdoptedStyleSheetsSet.call(this, sheets);
+        if (sheets.length > 0) {
+          patchHostConnectedCallback(this);
+        }
+      },
+    });
+  }
+}
