+++ vite.config.ts
@@ -35,13 +35,21 @@ export default defineConfig({
                 // the proper extensions will be added
                 fileName: 'css-anchor-positioning-fn',
               }
-            : // build that runs the polyfill on import
-              {
-                entry: resolve(__dirname, 'src/index.ts'),
-                name: 'CssAnchorPositioning',
-                // the proper extensions will be added
-                fileName: 'css-anchor-positioning',
-              },
+            : process.env.BUILD_SHADOW
+              ? // build that installs the constructed-stylesheet patches on import
+                {
+                  entry: resolve(__dirname, 'src/index-shadow.ts'),
+                  name: 'CssAnchorPositioningShadow',
+                  // the proper extensions will be added
+                  fileName: 'css-anchor-positioning-shadow',
+                }
+              : // build that runs the polyfill on import
+                {
+                  entry: resolve(__dirname, 'src/index.ts'),
+                  name: 'CssAnchorPositioning',
+                  // the proper extensions will be added
+                  fileName: 'css-anchor-positioning',
+                },
         emptyOutDir: false,
         target: 'es6',
         sourcemap: true,
@@ -77,7 +85,7 @@ export default defineConfig({
       provider: 'istanbul',
       reporter: ['text-summary', 'html'],
       include: ['src/**/*.{js,ts}'],
-      exclude: ['src/index.ts', 'src/index-fn.ts', 'src/index-wpt.ts'],
+      exclude: ['src/index.ts', 'src/index-fn.ts', 'src/index-shadow.ts', 'src/index-wpt.ts'],
       skipFull: true,
     },
   },
